fix cyclic dependency chatAction <-> chatEditorInput. fyi @roblourens
